add change and complete data provider for send to server

+++ kashka/Business Logic Layer/Models/Stuff_Code_Count_Pair.cs
@@ -16,7 +16,7 @@ namespace kashka.Business_Logic_Layer.Models
         /// <summary>
         /// تعداد/مقدار
         /// </summary>
-        public int Count { get; set; }
+        public decimal Count { get; set; }
 
         /// <summary>
         /// مبلغ واحد(ریال)
